BREAKING CHANGE: Refactor scan timing dataclass and methods to new scan_timing.py module.
